yatube_api/api/permissions.py
from rest_framework import mixins, filters, permissions, viewsets


class AuthorPermissions(permissions.BasePermission):
    def has_object_permission(self, request, view, obj):
        return (request.method in permissions.SAFE_METHODS
                or obj.author == request.user)


class CreateRetrieveViewSet(mixins.ListModelMixin,
                            mixins.CreateModelMixin,
                            viewsets.GenericViewSet):
    pass
